only do the cross check test if std is available for the target

+++ ci/script.sh
@@ -16,13 +16,15 @@ main() {
     export QEMU_STRACE=1
 
     # test `cross check`
-    td=$(mktemp -d)
-    cargo init --lib --name foo $td
-    pushd $td
-    echo '#![no_std]' > src/lib.rs
-    cross check --target $TARGET
-    popd
-    rm -rf $td
+    if [ -z $STD ]; then
+        td=$(mktemp -d)
+        cargo init --lib --name foo $td
+        pushd $td
+        echo '#![no_std]' > src/lib.rs
+        cross check --target $TARGET
+        popd
+        rm -rf $td
+    fi
 
     # `cross run` test for thumb targets
     case $TARGET in
